Convert functions b/w topological and triangle mesh
Remove standalone functions converting between topological mesh and
triangle mesh in favor of member functions

Added inline function file for topological mesh
